Added custom buffer sizes Added masking to Blit2D and BlitSprite2D Added define for console lines Added expirermental support for remembering last window position Fixed window size, now creates correct client size area Fixed retarded reference to resolved pointer (this) Moved UI class to its own project, seems useful for other applications

+++ neo/platformer.h
@@ -134,8 +134,8 @@ namespace NeoPlatformer{
 		Point<int> debugflags;
 
 		// Env geometrys
-		double dt;
-		double dtMulti;
+		float dt;
+		float dtMulti;
 		int tileWidth, tileHeight;
 		Rect<int> map;
 		Point<int> scroll;
